requirements.txt
numpy>=1.21.0
faiss-cpu>=1.7.4
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0
